requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
attrs==25.4.0
beautifulsoup4==4.14.3
black==26.1.0
boto3==1.42.30
botocore==1.42.30
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
lxml==6.0.2
multidict==6.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
numpy==2.4.1
packaging==25.0
pandas==2.3.3
pathspec==1.0.3
platformdirs==4.5.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.0
pytz==2025.2
RapidFuzz==3.14.3
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8.3
starlette==0.50.0
twilio==9.9.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
yarl==1.22.0
